app/contextlynx/requirements.txt
Django==5.1.1
gunicorn==23.0.0
django-environ==0.11.2
Markdown==3.7
numpy==1.26.4
scikit-learn==1.5.2
tokenizers==0.19.1
transformers==4.44.2
torch==2.4.1
psycopg2-binary==2.9.9
openai==1.45.0
spacy==3.7.6
beautifulsoup4==4.12.3
networkx==3.3
node2vec==0.5.0
pgvector==0.3.3
youtube-transcript-api==0.6.2